[oto.comoco.chart-history]:new: building api service 1/2 done
